requirements.txt
requests
numpy
opencv-python
pyserial
yfinance
mss
python-dateutil